+++ R/TernaryPlot.R
@@ -463,9 +463,7 @@ HorizontalGrid <- function (grid.lines = 10, grid.col='grey',
 #'        \code{\link[graphics]{points}},
 #'        \code{\link[graphics]{lines}} or
 #'        \code{\link[graphics]{text}}.
-#' @param coordinates A list, matrix, data.frame or vector in which each
-#'                    element (or row) specifies
-#'                    the three coordinates of a point in ternary space.
+#' @template coordinatesParam
 #' @param \dots Additional parameters to pass to \code{PlottingFunction}.  
 #' If using `TernaryText`, this will likely include the parameter `labels`,
 #' to specify the text to plot.