+++ man-roxygen/coordinatesParam.R
@@ -1,0 +1,3 @@
+#' @param coordinates A list, matrix, data.frame or vector in which each
+#'                    element (or row) specifies
+#'                    the three coordinates of a point in ternary space.