+++ NEWS.md
@@ -1,5 +1,6 @@
-# Ternary v1.0.2.9100 (development)
+# Ternary v1.0.2.9101 (development)
  - Add support for contour plots
+ - Rudimentary support for density shading
  - Maintain isometry when one of `xlim` or `ylim` is specified.
  - Fix issues with `TernaryPlot` documentation.
  - Remove names in `TernaryCoords`.